Hazards

+++ api/src/services/destroyGameByName.ts
@@ -3,9 +3,13 @@ import { Sequelize } from "sequelize"
 const destroyGameByName = (db: Sequelize) => async (name: string) => {
   await db.transaction(async (t) => {
     const game = (await db.models.Game.findOne({ where: { name } }))!
+    if (!game) {
+      return
+    }
 
     await db.models.Message.destroy({ where: { gameId: game.dataValues.id } })
     await db.models.Agent.destroy({ where: { gameId: game.dataValues.id } })
+    await db.models.Hazard.destroy({ where: { gameId: game.dataValues.id } })
     await db.models.Train.destroy({ where: { gameId: game.dataValues.id } })
     await db.models.Hop.destroy({ where: { gameId: game.dataValues.id } })
     await db.models.Line.destroy({ where: { gameId: game.dataValues.id } })
